Rename modelfinder package to weights
Rename FindModels to FindWeights
